move DeviceContext into it's own module
